Handle exception when user is not logged in (token absent)

+++ src/main/java/resource/PlaylistsResource.java
@@ -5,6 +5,7 @@ import jakarta.inject.Inject;
 import jakarta.ws.rs.*;
 import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
+import resource.exceptions.UserNotLoggedInException;
 import service.PlaylistsService;
 
 @Path("/playlists")
@@ -14,6 +15,7 @@ public class PlaylistsResource {
     @GET
     @Produces(MediaType.APPLICATION_JSON)
     public Response getPlaylists(@QueryParam("token") String token) {
+        if (token == null) throw new UserNotLoggedInException();
         return Response.ok().entity(playlistsService.getAllPlaylists(token)).build();
     }
 
@@ -21,6 +23,7 @@ public class PlaylistsResource {
     @Path("/{id}")
     @Produces(MediaType.APPLICATION_JSON)
     public Response deletePlaylist(@QueryParam("token") String token, @PathParam("id") String id) {
+        if (token == null) throw new UserNotLoggedInException();
         return Response.ok().entity(playlistsService.deletePlaylist(token, id)).build();
     }
 
@@ -28,6 +31,7 @@ public class PlaylistsResource {
     @Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.APPLICATION_JSON)
     public Response addPlaylist(@QueryParam("token") String token, Playlist playlist) {
+        if (token == null) throw new UserNotLoggedInException();
         return Response.ok().entity(playlistsService.addPlaylist(token, playlist)).build();
     }
 
@@ -36,6 +40,7 @@ public class PlaylistsResource {
     @Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.APPLICATION_JSON)
     public Response editPlaylist(@QueryParam("token") String token, @PathParam("id") String id, Playlist playlist) {
+        if (token == null) throw new UserNotLoggedInException();
         return Response.ok().entity(playlistsService.editPlaylist(token, id, playlist)).build();
     }
 
@@ -43,6 +48,7 @@ public class PlaylistsResource {
     @Path("/{id}/tracks")
     @Produces(MediaType.APPLICATION_JSON)
     public Response getTracksFromPlaylist(@QueryParam("token") String token, @PathParam("id") String id) {
+        if (token == null) throw new UserNotLoggedInException();
         return Response.ok().entity(playlistsService.getTracksFromPlaylist(token, id)).build();
     }
